+++ components/mdz/internal/domain/repository/portfolio.go
@@ -6,4 +6,6 @@ type Portfolio interface {
 	Create(organizationID, ledgerID string, inp mmodel.CreatePortfolioInput) (*mmodel.Portfolio, error)
 	Get(organizationID, ledgerID string, limit, page int) (*mmodel.Portfolios, error)
 	GetByID(organizationID, ledgerID, portfolioID string) (*mmodel.Portfolio, error)
+	Update(organizationID, ledgerID, portfolioID string, inp mmodel.UpdatePortfolioInput) (*mmodel.Portfolio, error)
+	Delete(organizationID, ledgerID, portfolioID string) error
 }