+++ components/mdz/pkg/cmd/portfolio/portfolio.go
@@ -14,6 +14,8 @@ func (f *factoryPortfolio) setCmds(cmd *cobra.Command) {
 	cmd.AddCommand(newCmdPortfolioCreate(newInjectFacCreate(f.factory)))
 	cmd.AddCommand(newCmdPortfolioList(newInjectFacList(f.factory)))
 	cmd.AddCommand(newCmdPortfolioDescribe(newInjectFacDescribe(f.factory)))
+	cmd.AddCommand(newCmdPortfolioUpdate(newInjectFacUpdate(f.factory)))
+	cmd.AddCommand(newCmdPortfolioDelete(newInjectFacDelete(f.factory)))
 }
 
 func NewCmdPortfolio(f *factory.Factory) *cobra.Command {